pyproject.toml
[tool.poetry]
name = "tasmota-power-monitoring-mqtt-to-influxdb"
version = "0.1.2"
description = ""
authors = ["Sven Grunewaldt <strayer@olle-orks.org>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.4"
influxdb = "^5.2"

[tool.poetry.dev-dependencies]
black = "19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
